[Chore] controller, service preq로 합침
