+++ src/main/java/kr/co/preq/domain/preq/controller/PreqController.java
@@ -2,13 +2,9 @@ package kr.co.preq.domain.preq.controller;
 
 import javax.validation.Valid;
 
-import io.github.flashvayne.chatgpt.service.ChatgptService;
-import kr.co.preq.domain.preq.service.ChatGptService;
+import kr.co.preq.domain.preq.dto.*;
 import org.springframework.web.bind.annotation.*;
 
-import kr.co.preq.domain.preq.dto.CoverLetterRequestDto;
-import kr.co.preq.domain.preq.dto.CoverLetterResponseDto;
-import kr.co.preq.domain.preq.dto.PreqResponseDto;
 import kr.co.preq.domain.preq.service.PreqService;
 import kr.co.preq.global.common.util.response.ApiResponse;
 import kr.co.preq.global.common.util.response.SuccessCode;
@@ -21,10 +17,6 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class PreqController {
 	private final PreqService preqService;
-	private final ChatGptService chatService;
-	private final ChatgptService chatgptService;
-
-
 
 	@PostMapping()
 	public ApiResponse<CoverLetterResponseDto> saveCoverLetter(@RequestBody @Valid CoverLetterRequestDto requestDto) {
@@ -41,5 +33,10 @@ public class PreqController {
 
 		//return ApiResponse.success(SuccessCode.PREQ_GET_SUCCESS, responseDto);
 		return null;
+	}
+
+	@PostMapping("/gpt")
+	public ChatGptResponseDto sendQuestion(@RequestBody QuestionRequestDto requestDto) {
+		return preqService.askQuestion(requestDto);
 	}
 }
