+++ src/main/java/kr/co/preq/domain/preq/dto/QuestionRequestDto.java
@@ -1,13 +1,9 @@
 package kr.co.preq.domain.preq.dto;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
-import java.time.LocalDate;
-import java.util.List;
 
 @Getter
 @NoArgsConstructor
